fix(design-system): separate LF and Apollo to avoid CSS conflicts on component DebugGrid
