+++ client/apollo/react/src/index.ts
@@ -16,7 +16,7 @@ export {
   type SpinnerVariants,
 } from "./Spinner/SpinnerApollo";
 export { Tag, type TagVariants, tagVariants } from "./Tag/TagApollo";
-export { DebugGrid } from "./Grid/DebugGrid";
+export { DebugGrid } from "./Grid/DebugGridApollo";
 export { ItemMessage } from "./Form/ItemMessage/ItemMessageApollo";
 export { ItemLabel } from "./Form/ItemLabel/ItemLabelApollo";
 export { TextArea } from "./Form/TextArea/TextAreaApollo";
